requirements.txt
flask
numpy
pandas
requests
tqdm
plotly.express